# [PERF] switch from realpathSync -> readlinkSync