@joliss has already identified nodes performance issues with realpath [here](https://github.com/joyent/node/issues/7902)

As per @krisselden's suggestion, we can likely just use `readlinkSync` here instead.

real life example: https://github.com/ember-cli/stress-app
before: `8668.970928ms`
after: `7152.094867ms`

`18%` improvement to incremental builds

As `realpathSync` uses `readlinkSync` internally, this also reduces calls to `readlinkSync`
before: `17826 (134.230726ms)`
after:   `6171 ( 93.061130ms)`