@joliss 2015-09-02
`readlink` just gives you the verbatim path of the symlink target, which may be relative to the symlink location. `realpath` is needed to turn it into an absolute path.

(One might be tempted to use `path.resolve` to resolve relative paths, but that isn't enough either, because symlinks *inside* the symlink path need to be resolved to deal with `..` correctly. For example, `/Users/stefanpenner/some_symlink` is no problem, but `/Users/stefanpenner/some_symlinked_directory/some_symlink` needs to resolve `some_symlinked_directory` first.)

I believe most of `realpathSync`'s overhead is in Node land, so if we could get a version that calls the POSIX `realpath(3)`, we might get a similar speedup.