@stefanpenner 2015-09-02
Should we just branch if the result is not absolute?

Short of a native extension (which causes portability issues) I'm not sure how to get posix version