@stefanpenner 2015-09-02
It may be an option to create a faster user land variation of realpathSync.

If you can provide a failing test case, it would be handy. That way I can be sure the scenario you describe is handled correctly