@stefanpenner 2015-09-02
Ya, I'm pretty sure we can implement a better then node's provided one, in user-space. In addition to that, if we took a similar approach as https://github.com/stefanpenner/mkdirp-bulk We could really cut down on duplicate internal stats and readlinks.