@joliss 2015-09-02
My preferred solution would be to fix Node to call `realpath(3)` on POSIX.