@joliss 2015-09-02
It's tempting to branch on absolute paths, but `path.isAbsolute` seems buggy on Windows (https://github.com/nodejs/node/issues/2656), so I'd prefer to put the proper fix.