@stefanpenner 2015-09-02
Fixing node seems fine, but that actually means months or years until our users gain access to it. I wish adding native extensions wasn't such a hazard.

Someone needs to bite the bullet an fix the node side of things. So future travelers don't have to endure this.

The more I think about the more I think the batching approach will likely yield the best short-term outcome for us. So I will likely explore both a faster user land variation and on top of it a batching version.