@stefanpenner 2015-09-03
@joliss i found https://gist.github.com/joliss/3292eda35cd4d564cdd7 i suspect this is the failing scenaro?