@joliss 2015-09-03
Yeah that's the scenario.