@stefanpenner 2015-11-23
> readlink just gives you the verbatim path of the symlink target, which may be relative to the symlink location. realpath is needed to turn it into an absolute path.

Does it sound reasonable for us to implement both? First use `readLink`, if the result of readLink is absolute, carry on. Otherwise fallback to `realpath`.

In our use-cases it is fairly uncommon to hit the relative scenario (although not impossible).