@joliss 2015-11-25
> Does it sound reasonable for us to implement both? First use `readLink`, if the result of readLink is absolute, carry on. Otherwise fallback to `realpath`.

Sounds very reasonable. `readlinkSync` + `isAbsolute` is more than 10x faster than `realpath` on my system, so this should give a good speedup.

If you want to give it a try, note that on Windows `isAbsolute` doesn't handle `\foo\bar` correctly (https://github.com/nodejs/node/issues/2656). (This matters when src and dest are on different drives.) We can maybe just use a regex to account for `\foo\bar` cases.