@stefanpenner 2015-11-25
Thanks for the Windows tip. I was not aware of that issue